backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
reportlab==4.0.4
Pillow==10.0.0
Werkzeug==2.3.7
